Entering and Exiting logic added
When a car enters it's stored in TimeTable table and when it exists the row will be transfered from TimeTable to the Transaction table and the check_out time will be added. This way the TimeTable table stays clean and results to faster search. Every requests puts the car in garage with id 1 by default, this will be changed later on.